Add BMI calculation logic and activity navigation

+++ app/src/main/java/com/fitapp/util/BMI.java
@@ -2,13 +2,13 @@ package com.fitapp.util;
 import java.io.Serializable;
 
 public class BMI implements Serializable{
-    public int kilo;
-    public int groesse;
-    public BMI  (int kilo, int groesse){
+    public double kilo;
+    public double groesse;
+    public BMI  (double kilo, double groesse){
 
 
     }
-    public double berechnen (int kilo, int groesse){
+    public double berechnen (double kilo, double groesse){
         double groesseM = groesse/100;
         double bmiresult = kilo/Math.pow(groesseM, 2);
